Fix printf format mismatches
Sponsored by:	The FreeBSD Foundation


git-svn-id: svn+ssh://svn.freebsd.org/base/head@264115 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/boot/amd64/efi/copy.c
@@ -51,8 +51,8 @@ x86_efi_copy_init(void)
 	status = BS->AllocatePages(AllocateAnyPages, EfiLoaderData,
 	    STAGE_PAGES, &staging);
 	if (EFI_ERROR(status)) {
-		printf("failed to allocate staging area: %d\n",
-		    status & EFI_ERROR_MASK);
+		printf("failed to allocate staging area: %lu\n",
+		    (unsigned long)(status & EFI_ERROR_MASK));
 		return (status);
 	}
